% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{get_dbl_scores}
\alias{get_dbl_scores}
\title{Computes admixture scores from}
\usage{
get_dbl_scores(
  score_params,
  df,
  so_rna,
  cell_type_adj_mat,
  get_sc_scores = FALSE
)
}
\arguments{
\item{score_params}{list of parameters including   signal_thres, exclude_genes,
p.c, min.expr.frac, exclude.cell.types, and max.prob.high}

\item{df}{Data Frame with transcript coordinates, gene, cell, and cell_type columns}

\item{so_rna}{Seurat object with scRNA-seq data for comparison}

\item{cell_type_adj_mat}{Cell-type adjecency matrix}

\item{get_sc_scores}{Logical whether to compute scores for the scRNA data as well}
}
\value{
a Data Frame of admixture scores for all cells
}
\description{
Computes admixture scores from
}
